Revert "fix upgrade and add testing"
This reverts commit 72fdbf6bfda1e6a5300656b24b56adf1507c45b5.
